feat: add layout components including Navbar, Hero, Matches, Voting, LineUp, and Footer; add voting banner image
